Store the start and end index of each bencode node, this will be useful later when we grab the info node and create an info hash

+++ include/bencode/BencodeDecoder.hpp
@@ -97,7 +97,7 @@ private:
     std::shared_ptr<std::vector<uint8_t>> buffer_;
 
     /// Position in the buffer
-    int pos_;
+    size_t pos_;
     
     /// A dictionary of decoded bencode data
     std::map<std::string, BencodeValue> info_dict_;
